requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bitsandbytes==0.43.1
blis==0.7.11
boto3==1.34.133
botocore==1.34.133
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cohere==5.5.8
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
DataProperty==1.0.1
datasets==2.20.0
dill==0.3.8
evaluate==0.4.2
fastavro==1.9.4
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
install==1.3.5
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lexicalrichness==0.5.1
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mbstrdecoder==1.1.3
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pathvalidate==3.2.0
peft==0.11.1
pillow==10.3.0
portalocker==2.10.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pyphen==0.15.0
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge_score==0.1.2
rsa==4.9
s3transfer==0.10.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
sqlitedict==2.1.0
srsly==2.4.8
sympy==1.12.1
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
termcolor==2.4.0
textblob==0.18.0.post0
textdescriptives==2.8.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
tqdm-multiprocess==0.0.11
transformers==4.41.2
triton==2.3.1
typepy==1.3.2
typer==0.12.3
types-requests==2.32.0.20240622
typing_extensions==4.12.2
tzdata==2024.1
uncertainty-calibration==0.1.4
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
word2number==1.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0
